add a few links, but mostly fix some bad smiles strings in the ORNL_FP.xlsx
There were 11 goofy smiles strings in the ORNL_FP.xlsx file that were not resolvable into molecules.

MTaylor hypothesized that these were all the same as the preceding molecule. I confirmed that the fingerprints and descriptors are the same as the preceding molecule.  I have replaced the bad entries with the SMILES string of that entry.

See 2023.10.18_checking_for_bad_smiles_in_ornl.nb